Create Task 1 Notes
Basic Concepts
Using certain in-built functions, such as bit_is_clear, we can see if the button is pressed or not, and use a counter variable to count the number of button presses within 5 seconds.
To accurately measure 5 seconds, we make use of the in-built timer, and by setting the required bits in the registers, we can make it count upto 5 seconds.
We also need to look into debouncing of the button as that can cause problems in increasing the counter variable. There are many methods of doing so, such as hardware debouncing using an RC circuit. Due to lack of capacitors, I have decided to use software debouncing. Basically, I’m using a counter, which will increase when the button is pressed. When it crosses a certain threshold value, it will increase the button presses counter. Hence, we are able to clearly see if the button correctly pressed or not.
Then we can display the result on the serial monitor using USART, and show the output on the LED’s by setting the required bits high corresponding to the number.