Sub Module for Communication Technology
Mahishmati is being attacked by Kalakeyan tribe. A monitoring system needs to be placed on the fortress in order to monitor the intruders into Mahismati.

So, design a tracking system to have a track of who's approaching the fortress and print the distance of the person from the fortress wall by spanning the tracker for an angle of 180° and beep an alarm when distance is less than 100 cm.